Upgrade dependencies
In order to prepare for merging into the cluster-etcd repository, which
builds with JDK25, we need to upgrade various dependencies (mostly
Lombok and Mockito). While we're doing that, we can upgrade a bunch
of dependencies to their latest versions:

spring-boot: 3.1.5 -> 3.5.8
Jackson: 2.15.2 -> 2.18.2
Logback: 1.4.11 -> 1.5.15
SLF4J: 2.0.7 -> 2.0.16
SnakeYAML: 1.33 -> 2.3
Micrometer Prometheus: 1.11.5 -> 1.14.3
JUnit: 5.10.0 -> 5.11.4
AssertJ: 3.24.2 -> 3.27.3
jetcd: 0.7.3 -> 0.8.6
Lombok: 1.18.30 -> 1.18.40

Some of the old versions had vulnerabilities, so upgrading them is
important.

Signed-off-by: Michael Froh <msfroh@apache.org>